+++ test/support/fixtures/notes_fixtures.ex
@@ -16,7 +16,7 @@ defmodule Memex.NotesFixtures do
       attrs
       |> Enum.into(%{
         content: "some content",
-        tags: [],
+        tags: ["example-tag"],
         slug: random_slug(),
         visibility: :private
       })