+++ changelog.md
@@ -1,5 +1,6 @@
 # v0.1.3
 - backlink to other notes in notes
+- search tags on click
 
 # v0.1.2
 - fix more typos